remove advanced mode button and functionality ( component not removed)
