+++ packages/editor/src/components/hierarchy/HierarchyPanelContainer.tsx
@@ -513,20 +513,18 @@ export default function HierarchyPanel({
     <>
       <div className={styles.panelContainer}>
         <div className={styles.dockableTabButtons}>
-          {editorState.advancedMode.value && (
-            <div style={{ flex: 1 }}>
-              {t('editor:hierarchy.lbl-explode')}
-              <Checkbox
-                className={styles.checkbox}
-                classes={{ checked: styles.checkboxChecked }}
-                value={editorState.showObject3DInHierarchy.value}
-                sx={{ marginLeft: '5px' }}
-                onChange={(e, value) =>
-                  dispatchAction(EditorAction.showObject3DInHierarchy({ showObject3DInHierarchy: value }))
-                }
-              />
-            </div>
-          )}
+          <div style={{ flex: 1, paddingLeft: '5px' }}>
+            {t('editor:hierarchy.lbl-explode')}
+            <Checkbox
+              className={styles.checkbox}
+              classes={{ checked: styles.checkboxChecked }}
+              value={editorState.showObject3DInHierarchy.value}
+              sx={{ marginLeft: '5px' }}
+              onChange={(e, value) =>
+                dispatchAction(EditorAction.showObject3DInHierarchy({ showObject3DInHierarchy: value }))
+              }
+            />
+          </div>
           <Search elementsName="hierarchy" handleInputChange={setSearchHierarchy} />
         </div>
         {Engine.instance.scene && (
