+++ packages/editor/src/components/styles.module.scss
@@ -100,20 +100,4 @@
 
 .dockableTab {
   display: flex;
-
-  .dockableTabButtons {
-    display: flex;
-    position: absolute;
-    width: 100%;
-    padding-left: 110px;
-    align-items: center;
-
-    button {
-      margin: auto;
-      background: none;
-      border: none;
-      display: flex;
-      align-items: center;
-    }
-  }
 }
