+++ packages/editor/src/components/toolbar/ToolButton.css
@@ -3,6 +3,8 @@
 .toolButton {
   width: 40px;
   height: 40px;
+  min-width: 40px;
+  min-height: 40px;
   border: none;
   color: var(--iconButtonColor);
   cursor: pointer;
